% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confDualLRTpartial.R
\name{confDualLRTpartial}
\alias{confDualLRTpartial}
\title{Dual Likelihood-ratio based Confidence Interval for Causal Effects}
\usage{
confDualLRTpartial(data, alpha = 0.05)
}
\arguments{
\item{data}{n x d matrix, first column X values (potential cause), second column Y values (potential effect)}

\item{alpha}{1-confidence level, defaults to 0.05}
}
\value{
A vector containing the bounds of the confidence region. First entry indicates if no effect is plausible.
}
\description{
This function calculates an asymptotic Confidence Interval for the Total Causal Effect based on the Dual LRT.
It assumes a Linear Structural Causal Model (d nodes) and additive Gaussian errors with partial equal variance
between the potential cause and effect.
It targets the Total Causal Effect of X (first column) on Y (second column) based on observational (sample size n).
}
