% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervalsDualLRTpartial.R
\name{intervalsDualLRTpartial}
\alias{intervalsDualLRTpartial}
\title{Compute the confidence intervals for the causal effect}
\usage{
intervalsDualLRTpartial(siginv, perm, L1, alpha, d, n)
}
\arguments{
\item{siginv}{Inverse covariance matrix of data}

\item{perm}{Set of causal orderings}

\item{L1}{Likelihood under alternative}

\item{alpha}{1-confidence level}

\item{d}{Number of variables}

\item{n}{Number of observations}
}
\value{
Vector of plausible causal effect intervals
}
\description{
This (helper) function computes the confidence region of the total causal effect for a set of causal orderings based on Dual LRT.
}
