% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderDualLRTpartial.R
\name{orderDualLRTpartial}
\alias{orderDualLRTpartial}
\title{Test if a causal ordering is plausible}
\usage{
orderDualLRTpartial(index, d, n, siginv, tresh, ord, crit)
}
\arguments{
\item{index}{Integer representing the causal ordering.}

\item{d}{Number of variables in the system.}

\item{n}{Sample size.}

\item{siginv}{Inverse of the covariance matrix.}

\item{tresh}{Likelihood ratio threshold.}

\item{ord}{ordering of potential cause and potential effect}

\item{crit}{Chi-square critical value for hypothesis testing.}
}
\value{
Vector of plausible causal orderings with corresponding likelihood value.
}
\description{
This (helper) function test whether a given causal ordering is plausible based on the dual likelihood ratio test.
Note that a causal ordering corresponds to a partition of the variables into three groups, before, between and after the cause and effect of interest.
}
